fix(slack): break modal routing import cycle
