+++ extensions/slack/src/monitor/events/interactions.modal.ts
@@ -3,28 +3,7 @@ import { dispatchSlackPluginInteractiveHandler } from "../../interactive-dispatc
 import { parseSlackModalPrivateMetadata } from "../../modal-metadata.js";
 import { authorizeSlackSystemEventSender } from "../auth.js";
 import type { SlackMonitorContext } from "../context.js";
-
-export type ModalInputSummary = {
-  blockId: string;
-  actionId: string;
-  actionType?: string;
-  inputKind?: "text" | "number" | "email" | "url" | "rich_text";
-  value?: string;
-  selectedValues?: string[];
-  selectedUsers?: string[];
-  selectedChannels?: string[];
-  selectedConversations?: string[];
-  selectedLabels?: string[];
-  selectedDate?: string;
-  selectedTime?: string;
-  selectedDateTime?: number;
-  inputValue?: string;
-  inputNumber?: number;
-  inputEmail?: string;
-  inputUrl?: string;
-  richTextValue?: unknown;
-  richTextPreview?: string;
-};
+import type { ModalInputSummary } from "./modal-input-summary.js";
 
 type SlackModalBody = {
   user?: { id?: string };
